Fix unsafe int64→int narrowing for MaxSnapshotsToKeep
Use int64(wouldKeep) instead of int(config.MaxSnapshotsToKeep) to
avoid potential truncation on 32-bit platforms (CodeQL high severity).

+++ weed/plugin/worker/iceberg_maintenance_handler.go
@@ -649,7 +649,7 @@ func (h *IcebergMaintenanceHandler) expireSnapshots(
 		for j := i - 1; j >= 0; j-- {
 			wouldKeep++
 		}
-		if wouldKeep <= int(config.MaxSnapshotsToKeep) {
+		if int64(wouldKeep) <= config.MaxSnapshotsToKeep {
 			// Would go below minimum, keep it
 			kept++
 			continue
